refactor: extract enable_auto_merge into shared github_utils module
- Move _enable_auto_merge to scripts/github_utils.py, import from both
  automerge_dependabot.py and post_major_review.py
- Add dep_name and new_version inputs to review workflow so run-name
  shows "Review {repo} PR #{n} major update to {pkg} {version}"
- Pass dep_name and new_version in dispatch step
