+++ .github/workflows/chore-automerge-dependabot-prs.yml
@@ -84,10 +84,14 @@ jobs:
         run: |
           echo "$MAJOR_BUMPS" | jq -c '.[]' | while read -r bump; do
             pr_number=$(echo "$bump" | jq -r '.pr_number')
-            echo "Dispatching review for PR #${pr_number}..."
+            dep_name=$(echo "$bump" | jq -r '.dep_name')
+            new_version=$(echo "$bump" | jq -r '.new_version')
+            echo "Dispatching review for PR #${pr_number} (${dep_name} ${new_version})..."
             gh workflow run chore-review-major-dependabot-update.yml \
               --field target_repo="$TARGET_REPO" \
               --field pr_number="$pr_number" \
+              --field dep_name="$dep_name" \
+              --field new_version="$new_version" \
               --field dry_run="${DRY_RUN}"
           done
         env:
